backend/requirements.txt
flask
python-dotenv
flask-cors
scikit-learn==1.6.1
numpy
pandas